LayoutTests/platform/chromium/fast/events/onunload-expected.txt

main frame - has 1 onunload handler(s)
CONSOLE MESSAGE: line 1: Blocked alert('unload') during unload.
You should have seen an unload alert appear.
